Change CirrusSearchElasticaWrite partitioning key
These jobs have been unable to keep up with the rate of execution
required. Help things along by defining multiple partitions per
cluster, with application support. The general goal is to engage
multiple pods in the execution of this job queue.

This intentionally leaves partition 0 unused. The problem is that this
partition is currently behind by 20 million messages. If we distribute
some messages into the new partitions and some messages into the
existing partition 0 then 1/3 of the updates will be backlogged while
the other 2/3 get applied immediately (assuming this increases
throughput as expected).

Depends-On: I1496b915c7a90a493af3c11d4bafb96f4cc3a844
Bug: T314426
Change-Id: I09e658e7945d8d6eb3fb454216524e28220b77f9