api-gateway: disable shipping logs to eventgate
Since disabling debug logging (which was creating far too much output
and affecting logstash), we've lost the ability to accurately
investigate requests in the logs and we get very little output to
stdout or in logstash. Disable shipping logs to eventgate for the
short term while we investigate outputting the log twice.

Change-Id: I297ed7d907df1b6909bb5009e43afa6195d824ec