Add a helmfile configuration for the dse-k8s-eqiad cluster
This allows helmfile to know where to find the values files for the new
dse-k8s-eqiad cluster. We also create a cluster-group values file,
although there is only one cluster in the group for now.

Values relating to knative and kserve have been commented out, since
we may want to install these components in the near future.

Bug: T310196
Change-Id: Ie2d9315a25bf33530647426ba3b510c3b08733ba