api-gateway: Distinguish between internal host and host header setting
Previously the under-undocumented "host" parameter was added for
discovery hosts which allowed the setting of a custom internal
hostname to simplify the naming of discovery entries. This
accidentally collided with our recently added "host" paramter for
setting custom host headers for APIs before making the query to the
internal service. Distinguish between the two by renaming the internal
DNS hostname parameter "internal_host".

Change-Id: I89dc9fe4056371c3995710ad003ee31447a89e4c