Attempt to run the MCE and MAE consumers in the GMS container
It is possible to run the MCE and MAE consumers within the context of
the MS process, but in order to do so we need to modify our datahub helm
chart to facilitate this.

This change uses a global value for datahub_standalone_consumers_enabled
to determine whether or not to set two new environment variable in the
GMS pod.

We have been advised to try running these containers in non-standalone
mode in order to try to troubleshoot a specific ingestion problem.

Bug: T317053
Change-Id: I693d70d98d0c0e7d354edf3b61a64d4d0f73389c