admin_ng: set more values for Istio DR in ml-serve clusters
At the moment the envoy sidecar shows this config for the allowed
endpoints outside the mesh:

"max_connections": 100,
"max_pending_requests": 4294967295,
"max_requests": 4294967295,
"max_retries": 4294967295

The defaults seem to be not really ideal for us, so this change
adds a little bit more restriction. The max_requests field is
particularly important since we go through load balanced services,
and we want to avoid sticking with specific TCP conns for too long.

Bug: T313915
Change-Id: I953ab5574a026ad36eb7a587a48a72de98f208e5