admin_ng: use TLS origination for ml-serve sidecar configs
While trying to debug a perf issue with the MW API,
I realized that the Istio sidecars on ml-serve don't report
metrics for endpoints like api-ro.discovery.wmnet.
After some research, this turned out to be expected because
we originate a TLS connection from the Python code running
the model server, and the sidecar pod can only produce TCP-level
metrics. Istio offers "Egress TLS origination" to get more telemetry,
namely issuing a request via HTTP and then turning it to HTTPS
in the sidecar. The configuration depicted in [1] seem to
suggest that a VirtualService setting is not needed, but after
some tests (and after seeing the config reported in
https://github.com/istio/istio/issues/33105) my impression is that
only newer versions of Istio can do that (so basically having to
set only a DR and a SE, without any VS config).

Tested this config manually on ml-staging-codfw, it works and
I can see some new metrics in the Istio dashboard in Grafana.

For the moment let's move only api-ro.discovery.wmnet's configuration.

[1] https://istio.io/latest/docs/tasks/traffic-management/egress/egress-tls-origination/

Bug: T313915
Change-Id: I07bf8977129e5dc7c8b042af0c53cd141bc16527