kserve-inference: refactor rendering of isvc configs and add dnsConfig
Mapped values cannot be rendered in the current isvc config files,
because we are forcing a name: value pattern. Using a more permissive
toYaml simplifies the complexity of the rendering code and
allows to set values like dnsConfig.

The dnsConfig setting is passed to Knative that provides to add it to the Pod
resources created for new revisions (tested manually in staging).
This is needed to mitigate the issue indicated in the Phabricator bug,
namely too many DNS queries to coredns. The default ndots is 5,
and we want to lower it down to 3.

Due to how the isvcs are rendered (transformer optional), we can only
set a default for the predictors.

Bug: T313915
Change-Id: Ic76ba604685701d03bbac641b4f7f0ebd5a61827