coredns: add rewrite actions to the config map
In the ML clusters there is the need to tune some DNS
records to avoid short TTLs and bursts of DNS queries
to the coredns pods.
The large requests volume comes from Istio/Envoy,
since most of the Knative svc endpoints are defined
as Envoy clusters with STRICT_DNS, and hence Envoy
follows the DNS TTL set for the record refreshing
its IP cache accordingly.
The vast majority of the ml-serve DNS traffic (from Envoy)
is for cluster-local-gateway.istio-system.svc.cluster.local.,
that is defined as a regular Kubernetes svc ClusterIP.
CoreDNS seems to set 5s of DNS TTL for records if not specified,
that fits the problem really well. It offers rewrite actions
(see https://coredns.io/plugins/rewrite/#ttl-field-rewrites),
so this is a proposal to explicitly allow these kind of actions
in coredns' config.

Bug: T318814
Change-Id: I684d3bcbf0197925e5ad575fabbc40168a46230a