admin_ng: add custom DNS ttl rewrites for ml-serve clusters
Due to the default DNS TTL of 5s set by Envoy, most of
the Knative/Istio records are fetched almost continuously
from all the pods in the Istio Mesh to refresh their
IP list. This causes a huge amount of traffic to hit coredns,
that in turn slows down and causes high latencies for even
a simple load test of ML pods/traffic.

Ideally we should put this fix in Istio/Envoy, but as stopgap
measure (while we find how to do it and collaborate with
upstream) we want to increase the 5s TTL to 30s. The setting
has been applied manually to one of our cluster (~170 pods),
reducing the DNS traffic volume to coredns from ~5k rps to less
than 1k rps.

Bug: T318814
Change-Id: I4586bda84f4fb65a0d8a54adb4f9ebf36a166e7b