admin_ng: set proper TLS egress origination settings for ml-serve
The Istio sidecars can either let HTTPS/TLS connections through,
or wrap HTTP requests in a TLS tunnel transparently.

The way it works is described in:
https://istio.io/v1.9/docs/tasks/traffic-management/egress/egress-tls-origination/#tls-origination-for-egress-traffic

Essentially, TLS is mentioned in the Destination Rule
and the Service Entry config hints the sidecar about
how to upgrade from HTTP to HTTPS (for example, what
port to use etc..).

This change refactors the configs for api-ro.discovery.wmnet
to be more consistent with what upstream suggests,
and adds the TLS egress configuration for
eventgate-main.discovery.wmnet.

Change-Id: Ibbb86128e32c89deb25040246755b13b9bfd00f4