admin_ng: set x-forwarded-proto for ml-serve TLS egress orig settings
The MW API is currently returning a warning in every response to Lift
Wing due to the Istio TLS egress origination settings.
The x-forwarded-proto seems to be set to 'http', before it gets
wrapped into a TLS session to api-ro. This patch explicitly
sets the x-forwarded-proto to the correct value.

Bug: T320374
Change-Id: Ifdb7974d3162e8bbc1d00f9a95002b6d18bca058