calico: More calico 3.23.3 additions.
Typha does no longer ship with dedicated RBAC rules and uses the
calico-node service account instead. Also the docs do not include all
RBAC rules required for calico-node to work (but the upstram chart does
- almost).

Felix does assume different defaults for ipipMode and vxlanMode in
IPPools (at least it seems like that). Create IPPools with both
forcefully disabled for now.

Disable management of the kubeconfig file/token for the CNI binary in
calico-node. That file is provided by puppet and the calico-node service
account lacks permissions to update that token.

Bug: T307943
Change-Id: I2ed886ef707cfa5b97c08932d2808049a90a1baa