Add a spark-operator chart and helmfile configuration
This allows us to deploy the spark-operator to the dse-k8s cluster.
It is part of the admin_ng structure, since we require some cluster-wide
permissions and should therefore only be deployed by an SRE.

We limit the action of the spark-operator so that it only has the
privileges that it requires to manage SparkApplications and
SparkScheduledApplications in a single namespace called spark.

The operator itself runs in a namespace called spark-operator, but
monitors the spark namespace and creates driver pods in response to the
relevant API calls. The driver pods themselves subsequently create a
number of executor pods, again within the spark namespace, to run the
application.

In the production environment, we create NetworkPolicy objects
permitting the pods in the spark-operator namespaceto access the
Kubernetes API. We also create similar objects permitting the
spark-driver pods in the spark namespace to access the Kubernetes API.

There are two CustomResourceDefinition files included with the chart.
The operator itself is capable of adding these, but our configuration
disables this feature and adds them with helmfile instead. This
deployment does not declare any of the custom resources, so the CRDs may
be deployed in the same helmfile deployment as the rest of the chart.

We grant the rights for the managed spark-deploy infrastructure user to
be able to manipulate SparkApplication and ScheduledSparkApplication
objects in the spark namespace. This is the way that users are intended
to launch spark jobs.

Bug: T318926
Bug: T322635
Change-Id: I679ad5871d26b58af139100ffe15ec7f94eab070