Add basic rate-limit capabilities to ML clusters
The idea is to add a basic configuration option for the Istio mesh
proxies to apply local rate limit filters on incoming traffic.

The use case that we have in mind is to protect our inference API
from Hadoop/Spark/etc.. jobs not tuned correctly (so potentially
hitting the API with a big traffic volume).

The filter is currently a coarse grain one: all the clients are
considered equal, no priorities or specific settings applied
(only a ceiling on how much traffic a single pod can sustain at peak).
In the future we will probably change this setting, but we'd like
to start from somewhere to get experience with it.

Add also local rate_limit metrics to the Istio config (sadly
it will require a restart of all the pods to refresh the envoy
config).

Bug: T300259
Change-Id: I8383159a4711f5532c389b76bb51936a08f72d3a