knative-serving: improve chart's dependencies
The 'dependecies' typo in Chart.yaml masked a faulty
dependency tracking/implementation for a long time.

We could try to use 'helm dependency' to manage them,
but it may lead to traps and unexpected results. For
example, in this case if we add a charts/ subdir
with knative-serving-crds's .tgz file (so explicitly
adding the dependency following what Helm suggests)
we'll end up re-deploying the CRDs via helmfile,
probably ending up in an error.

The approach proposed is very simple: just document
the dependency without any explicity tie between them,
and deploy them in a safe order via helmfile procedures.

Bug: T303279
Change-Id: I50ad912ae8398a00d72d5cff9ffd6bc522bd7e8d