admin_ng: set thumbor max memory limit higher
In order to cope with surges for things like 3d2png calls and
gifsicle, increase the pod memory limit for thumbor. This limit will
be used for spikes in load but the request will be significantly
lower. This limit is intended for multiple instances in a pod rather
than a single instance using this amount constantly.

Bug: T233196
Change-Id: Id1c421ccc4d3f8ddc70341e8384e50b64363a226