api-gateway: Correct behaviour when handling multiple regexes in path
Previously having a regex in the path would cause the substutition of
paths to break because the first group was always assumed to be the
rest-of-path group. Add a debug model to test this.

Generally be more permissive in matching as far as slashes.

Fix internal_host values in lw models.

Bug: T317326
Change-Id: I8fc13e878ecc9d0f21bf667a2de88222322bf8e3