modules/mesh - Support mesh service proxy without exposing a Service for public_port
This allows us to use the the service mesh for egress,
without exposing a listener if it isn't needed.

So bump a minor version for the following modules:
* mesh.configuration
* mesh.deployment
* mesh.networkpolicy
* mesh.service

Please note: we're updating the dependency in mesh.name 1.0 so any
future change to modules in a chart using the mesh module will
require an upgrade to 1.1, which is the desired effect.

Also, update the scaffold and add a fixture.

Bug: T326252
Change-Id: Ifc46cc108e45481b447776e1b0d08764252e4dfe