Add flink-app-example service in the dse-k8s-eqiad cluster
Will use this for the first test usages of the flink-app chart.

This app uses pyflink and the Table API datagen connector
to print 2 random table row values per second to stdout.

This is the first 'service' to be deployed to dse-k8s-eqiad.
Add a new helmfile.d/dse-services directory to contain
service helmfiles deployed there.

Bug: T324576
Change-Id: I091223f6b53b0925043b30509956e2cde0d88ccc