changeprop: add liftwing revscoring streams
The ML team is trying to replace ORES with LiftWing,
and part of the effort is to replace
the mediawiki.revision-score stream with multiple
(more granular) ones. Changeprop is the engine
that powers revision-score, since the events
are generated after calling ORES in sys/ores_updates.js:
- Changeprop reads events from the mediawiki.revision-create
  topic (correspondent to new edits).
- Then it calls ORES' /precache endpoint, that returns
  multiple scores (one for each model compatible with
  a given wiki) in a single response.
- Then it adds all the scores in a revision-score json
  event and sends it to EventGate.
We want to replace the above with a simpler approach:
Changeprop holds different configurations, one for
each model type offered by Lift Wing. In every config:
- Changeprop listens to mediawiki.revision-create
  events.
- Based on a simple regex (the models that we offer
  don't support all wikis by default) it decides
  when to make an HTTP POST call to Lift Wing,
  adding the revision-create event as payload.
- Lift Wing's model server is configured to send
  an event to EventGate.

This change is meant to provide a simple/default
config for all the model-servers, flexible enough
to avoid (hopefully) too many changes to the chart
(and rely only on helmfile's tweaks).

Bug: T327302
Change-Id: Ib5b7d62017b4a5ce63ab8d5cad0d89ebd3922263