services: change liftwing's test rules
There seem to be no action taken by changeprop when a
mediawiki.revision-create event is created, despite metrics
indicating that the rule is evaluated. In this change
I want to relax the match rule to see if the regex is at
fault, and at the same time change the source kafka topic
to avoid any unwanted impact triggered by other rules
(that have production values in their config).

Bug: T327302
Change-Id: Id2b83d331b5b94f7995d0948c636f6a68ebfde8e