changeprop: update liftwing's rule templating
So far I didn't manage to see an HTTP request on LiftWing,
meanwhile I do see the related rule being processed by
Changeprop. The only big difference with the rest of
the config is that other rule with a single match/exec
clause do not have any "cases" lists.
Improving also the retry mechanism to be resilient on 503,
rather than ignoring them.

Bug: T327302
Change-Id: I2ed2b59257d47376add595da7cbfd4f0a6fe0ced