changeprop: refactor match template for liftwing
While writing the documentation for the new rules I realized
that the match config should be more flexible, since people
may need to match different fields depending on the topic
that provides the events (with a certain schema).

The change should be a no-op. I tried to use toYaml
to be more flexible, but I ended up stumbling on
https://github.com/helm/helm/issues/4262 and we'd need
to have quotes around regexes. As a workaround
I implemented more specific range loops for the two
use cases that we have (meta.field or field).

Bug: T327302
Change-Id: Ice8cd9ff1635a914422ccdceeeed8a7d3998cf62