Include wikivoyage in page/html rerenders
Why:
Apparently we 've been limiting page/html (aka mobile-sections)
rerenders to just wikipedias. Relevant change was merged in 2017

https://gerrit.wikimedia.org/r/c/mediawiki/services/change-propagation/deploy/+/373336

We want to rerender for other projects too of course since other
projects now use mobile-sections as well.

What:
Add wikivoyage in the regexp. This is a first step to see if the process
will fix T226931, we 'll have to evaluate the approach and figure out
next steps from there

Bug: T226931
Change-Id: Ifa74e9eeeccf27c8c9e736e59548112a6c53c25b