Include wiktionary in page/html rerenders
Why:
Apparently we 've been limiting page/html (aka mobile-sections)
rerenders to just wikipedias. Relevant change was merged in 2017

https://gerrit.wikimedia.org/r/c/mediawiki/services/change-propagation/deploy/+/373336

We want to rerender for other projects too now, since apparently there
are projects that use mobile-sections as well

What:
Add wiktionary in the regexp. In 416a37b1db, we added wikivoyage and it
worked, so this is expected to work as well.

The main issue that we face here is that we don't have a good list of
projects enable this for and the RESTBase deprecation plans indicate
that it would be a bad investment of time to allocate a lot of time to
find out. Hence, we will be adding them slowly and sparringly on a case
by case basis

Bug: T226931
Change-Id: Ic7629a4b52ae9f675628e6c73fdb20e7039bebf3