Run kubeconform on supported versions of charts & envs
Prior to this change the CI tests did not introspect the charts to
determine which Kubernetes versions they are compatible with. In
addition CI ran the kubeconform environment tests for all versions,
regardless of the version running in the environment.

After this commit we run kubeconform only for the Kubernetes versions
which are compatible for a given chart. In addition we determine the
configured Kubernetes version for an environment and run
kubeconform only for that version.

The impetus for this change was the vendoring of the jaeger chart which
depended on Kubernetes features not present in 1.16, which caused CI to
error out.

In order to compare versions github.com/davidrjonas/semver-cli was added
to the docker image. There was not another obvious way to incorporate
all the logic from github.com/Masterminds/semver, which is used by helm.

This commit also moves all test fixtures into
.rake_modules/spec/fixtures so there are no dependencies on repo
contents itself.

Bug: T320554
Depends-On: I8ff41f1a2f9d0a5b07a674a55df032c119ddee0b
Change-Id: I77657a2674a4546aa5088660745f09eedd5d2201