istio wikikube: Add the proper tolerations
Why:
ingressgateway Daemonset does not have the proper tolerations in
wikikube to allow it to be schedule on kask dedicated nodes. This in
turn causes the service, which is of type

 External Traffic Policy:  Local

to not work as there is no local pod to serve the traffic, failing
health checks

What:
Add 3 tolerations. NoExecute, NoSchedule and CriticalAddonsOnly to make
sure the pods get deployed

Change-Id: I108adca7eda7a48204463525db96606e56b2444e