services: tweak lift wing endpoints to allow wikidata-specific endpoints
The Wikidata model servers for article topic and quality are called
"item" topic and quality, and at the moment they are not reachable.

The simplest way to proceed is to relax the regex/pattern-match
in the API gateway to support both.

We'll need to document this in the API portal wiki as well,
but it should be easy.

Change-Id: Ic19d8b64ea576e8cc66f85cd6500bc840c978858