Bump datahub version to 0.10.0 and re-enable standalone consumers
We have built containers for version 0.10.0 of DataHub and wish to
deploy them. At the same time, we will be re-enabling the standalone
mae-consumer and mce-consumer pods, rather than have them running in the
context of the GMS. We tried this configuration recently for
troubleshooting, but now wish to revert.

Bug: T329514
Change-Id: Id0d45597086b55c13c3916c10047917fd8ececbc