admin_ng: set 'deploy' for ores-legacy in ml-serve's config
The ML team wants to deploy the ores-legacy service without
following the "kserve" way, namely not using the istio-proxy
sidecar and its features. The idea is to be able, on paper,
to switch the service from ml-serve to wikikube clusters
transparently.

The 'deploy-kserve' role doesn't have permissions to manage
stuff like istio configs, so it makes the ores-legacy deployment
to fail. We should use the 'deploy' user instead.

Bug: T330414
Change-Id: If629c50ad12f4e52425dc8d37a5e7b5da562ba01