modules: allow istio gateways to have more selectors (part 1)
The ML team is trying to experiment with multiple istio gateways
to separate concerns between kserve and the rest of the services
deployed (like ores-legacy etc..).

This is the first part of the change, in which the istio 1.0.0
template is duplicated as it is to 1.0.1 to ease the review process.

Change-Id: Ia021d623a65254eff5dbcaee0894b1a91146d54d