Rake: run with heuristics by default
This change introduces the ability, by default, to run only the tests we
need and not every check on every change.

It uses relatively naive heuristics but should cover 99% of cases, while
saving about 90% or more time for each commit.

In this change, I showcase the ability to track dependencies by making
a spurious change to the kafka-dev chart which is a dependency of
multiple charts which have on their own multiple deployments.

The time to run all the needed checks goes, on my computer, from  4
minutes to 35 seconds. We can hope in a similar or better gain for the
average change that touches only a few charts/deployments.

Change-Id: I540bab04a06e5af4eda79342e2f0fc706e23c6c7