machinetranslation: Remove args, document env vars
Why:
Become more consistent with the other aspects of how we configure the
software which is with environment variables

What:
Remove args from the chart, add a listing of the default values under
config.public.

Set config.public for our production env.

Bump to the image version that actually supports this

Bug: T331505
Change-Id: I6a4f83a5818e2429278f1c078e52451bb245a806