thumbor: haproxy timeout changes, block /metrics
Change our server timeout to be in keeping with what it should be in
line with the docs.

Block requests to /metrics on thumbor workers.

Healthcheck thumbor workers more often and be less strict about what
makes a node healthy.

Add timeouts to the stats backend to quieten log spam.

Bug: T334488
Change-Id: I96d9600b6f1ffdf282dc9d297f89d03e25a10b2f