flink-session-cluster: enable rocksdb metrics and increase jvm heap
the wikidata job running in codfw is struggling to load its checkpoint.
The size of the checkpoint seems particularly large (6.5G) compared to
its equivalent in eqiad (1.6G).
Enable rocksdb metrics to measure if anything weird is happening in the
number of keys and increase the heap given to the taskmanager to help
to load this state.

Bug: T336134
Change-Id: Ia147df0662ae5068939f6161d0f3b47978322a74