helmfile.d: add revertrisk model config to ml-serve clusters
This only adds the revertrisk-multilingual and -language-agnostic
models, as -wikidata is still experimental. It also removes the two
mentioned rr models from the experimental namespace. A different change
(919364) will add the -wikidata model.

This change also moves to using the non-experimental S3 bucket for model
storage. Note that the URL change subtly in this regard: /revertrisk/
becomes /revertrisk-language-agnostic/.

Bug: T333124
Change-Id: Ie07a0c918a5b83e2505192daca41c303e4796ecb