flink-operator custom basic egress
We need to allow the operator to talk to Zookeeper for managing JobManager HA.
The default egress network policy is to deny all.  This change adds
WMF's vendor/base/networkpolicy_1.0.0.tpl, and enables using it
using the usual values egress.enabled and networkpolicy.egress.dst_nets/dst_ports.

Because this change only adds templates to flink-kubernetes-operator chart,
it does not modify any of the upstream provided templates, it is easy enough
to maintain.

Add flink-operator egress to Zookeeper main-eqiad in DSE,
and Zookeeper test-eqiad in staging-eqiad.

Bug: T331283

Change-Id: Iafc8648fcf6325f9112875cb76324a4edd9cfbdd