kserve-inference: use dict instead of lists for inference_services
We have two use cases for isvcs:
- revscoring-related ones, in which we have a different number of pods
  in staging and production since we deploy less isvcs in staging
  (for resource constraints reason).
- all the other isvcs, that we deploy to both staging and prod
  and we only tweak their settings.

List cannot be overridden in helmfile, so we'll have difficult
times when trying to override values for the various environments
in the second use case. For example, we want to avoid autoscaling
settings that we set for prod clusters in staging, and having dicts
in there should help.

The revscoring ones shouldn't be touched since we'll deprecate them.

The end goal is to be able to override settings like outlined in
the helmfile revert-risk's staging config (without big copy/paste
actions).

Change-Id: Ic7bca80f8ed457ba18d2a3824fc15a20d6166097