mw-page-content-change-enrich - enable upgradeMode: savepoint, and take periodic savepoints.
updgradeMode: savepoint will cause a savepoint to be taken when restarting/upgrdaing the job,
and that savepoint to be used to restore state (kafka offsets).  Currently,
in stateless mode, the kafka offsets committed to kafka are being used, which is fine, but not
fault tolerant from Flink, since Flink's offset commit mechanism is meant to be used for monitoring,
not job recovery.

Periodic savepoints will allow us to recover from a recent savepoint if the JobManager and job
unexpectedly go offline.  If this happens, we'll have to supply the savepoint manually via
the initalSavepointPath config.

When we enabler HA, we can consider using upgradeMode: last-state instead.

Bug: T325303
Change-Id: Ib08fdbc304b063a8595fae155a9a2fba1380a00d