admin_ng: add the ml-serve experimental namespace to production
The experimental namespace is useful to test new ML model servers
during the early phases of its development. It is also good to make
generic experiments on a different hardware, like for example GPUs.

The ML team needs to experiment GPUs on ml-serve1001, running LLMs
on it, so this change adds the possibility to do it.

Bug: T334583
Change-Id: I3dfe45062bbe6019be300c51c4edba394ca03b58