mediawiki: Test sleeping before draining in envoy
On pod deletion, the SIGTERM signal to the pod is sent simultaneously to
the endpoint update. Envoy summarily closes connections on SIGTERM.
This results in a race condition where requests are still being sent to a pod
that is no longer accepting connections.
By sleeping for half of terminationGracePeriodSeconds before draining
envoy connection, we hope to delay connection draining enough that no
new connections are being sent to the pod and we can cleanly start
draining existing connections.

Bug: T331609
Change-Id: Ib0d5a287d298179b4d4d43c0cbc51623160a8bb0