kserve-inference: refactor the predictor's container settings
We only had the need to set the "resources" predictor container settings
up to now, but with LLMs we'd also need to set up different
things like the readinessProbe timings.

This change modifies the chart's templates to use a new key called
"container" in which we specify what we need to be rendered
in the isvc's predictor container settings.

Bug: T334583
Change-Id: Ie1c61474f49612c3461e192037ca6a10ea997611