ml-services: correctly override values in staging
Explicitly define "amd.com/gpu: 0" for staging since we don't have
any GPUs on ml-staging-codfw. Since helmfile value files are merged, if we don't
explicitly override this value then the result is merged, resulting
in an unspecified state.
The alternative would be to rollback to providing lists instead of
dicts to avoid this unwanted behavior.

Bug: T334583
Change-Id: Iada368fbe69e2e171c6f5b6c4f62ca7b8f0cd223