Enable the networkpolicy for datahub batch jobs
We are creating several new batch jobs to facilitate datahub upgrades
and maintenance. These require network access to the back-end data
stores. Previously we had a condition that only enables the relevant
networkpolicy object if the setup jobs were enabled, which they are not.

This commit updates the condition to enable the networkpolicy if the new
batch jobs are enabled.

Bug: T329514
Change-Id: Ie2f562d41a86de0aec1f626648df42f4dd3c339f