Fix the networkpolicy selector for datahub maintenance jobs
The main datahub chart creates a number of batch job objects, but the
labels on the pod templates that they created were not correctly matched
by the networkpolicy.

This commit adds a Values.global.podLabels label, which is automatically
applied by the batch job pod spec templates. It will cause the
networkpolicy to be applied to the resulting pod objects.

Bug: T329514
Change-Id: I73538b0d48f72333dcf54a03acaaff3d86402490