Specify the schema registry type for datahub
Recent versions of datahub have mitigated the dependency on an external
schema registry by enabling the option of an internal registry.

However, we have already deployed a version with an external registry
based on Karapace. As such, we have to specify which type we are using,
or some required environment variables are not made available to the
containers.

Bug: T329514
Change-Id: Iee8abcd5538508e36df4c12dc3cd740bc74be266