Permit datahub batch jobs to contact the GMS service
Previously we had a networkpolicy object that permitted all of the other
datahub applications to contact the GMS service. This didn't work when
we started to enable the batch jobs that perform maintenance tasks,
since they don't have an 'app' label.

This change permits this batch jobs, which have a 'workload: datahub'
label to contact the GMS as well, facilitating upgrade and other
maintenance work.

Bug: T329514
Change-Id: I937dc6754b48b4525add5486812609b5b35bb0ce