Enable the datahub setup jobs for mysql and elasticsearch
These batch jobs are part of the maintenance and initialization of
datahub, which happens at deployment time. At present we have not used
them, but we now wish to enable them during the deployment to staging.

We add the 'workload: databub' to the job pod template, in order that
the existing networkpolicy for the top-level chart is applied to them.

Bug: T329514
Change-Id: Idd5194bca5d07fec439430e6c808dd23ea0dc594