Disable the datahub upgrade job
There are two jobs that run prior to the main datahub tasks. One is
called datahub-main-system-update-job and this is working. The other is
called datahub-main-nocode-migration-job but this is not working. We
would like to disable it for now, so that we can complete a deployment
to staging and continue testing.

Bug: T329514
Change-Id: Ic550c5a14c6454b61c3d1b8ea268593916dcd815