Enable the kafka-setup job for datahub in staging.
We have previoulsy not used this job but now wish to do so, starting
with our deployment in staging.

In addition to enabling the flag to run, we add a few more environment
variables and parameters to enable them.

Finally, we also deploy the updated datahub images which contain an
updated script to setup the kafka topics.

Bug: T329514
Change-Id: Ie480dfeee380ec146f2d242700088453a6eeb077