modules: Add a new networkpolicy for base modules
Why:
We want to somewhat automate which database sections we will be allowing access
to. Per T340843 automating nicely and fully this isn't feasible, so we
need to go with a stopgap. For now the stopgap is a "contract" with Data
Persistence, where we allow access to specific ports. While we 'd prefer
to also have stablic virtual IPs in the contract, this isn't currently
possible.

What:
Generalize the approach we took with MediaWiki chart, where we just
opened access to the entire 10.0.0.0/8 to ports 3310-3320 and 3306.

Introduce a new define that utilizes a new stanza of data called
mariadb. In production, a hash named mariadb.section_ports will exist
that contains the mappings of all sections to their equivalent TCP
ports. This will get populated by puppet, see the Depends-on change at
the end of this commit

We then allow each chart to define the sections it wants access to and
open the respective egress ports. 3306 is always opened if a chart
defines any section as it is the same for all sections.

Bug: T340843
Bug: T341117
Depends-on: Id6172c7cbd8e13dbe6571ea0d91c6c80ba89b0ce
Change-Id: Idf9d64a026ed33c170066d7c84af8c9d366d7ada