add Benthos chart
Benthos will be helpful for multiple data processing use cases, for a start for
replacing RESTbase cache invalidation logic. This adds a generic Benthos chart
to be used by specific services. Note that this is WIP and not complete -- in
particular, a way to meaningfully configure Benthos is not present yet.

Bug: T324200
Change-Id: I7d916342f3df508adb6f2c57791a65538810c7fd