Bump the image of datahub to the new 0.10.4 containers
This is the first version of DataHub that is built from pristone
upstream sources, rather than from our own fork of the upstream
codebase. It upgrades DataHub itself to version 0.10.4.

We disable the kafka-setup job because of an error generated during the
last part of the script, which we are still troubleshooting.

Bug: T329514
Change-Id: I83f5fb319ae5ab1921898243cc7ba2872246638b