Fix the datahub frontend authentication
The newest version of the datahub-frontend container sets JAVA_OPTS
explicitly before running the start.sh script. This invalidates our
previous method of using JAVA_OPTS to supply the location of our custom
JAAS configuraiton file.

Therefore we have to change the way that we provide this and overwrite
the /datahub-frontend/conf/jaas.conf file itself.

Bug: T329514
Change-Id: I544166c222fc65f627afcf7e1278cc39303f76a9