services: allow kafka batches in EventGate's main producer
Most of the EventGate messages/events are sent with the guaranteed
producer, that currently sets queue.buffering.max.ms=0 (alias for
linger.ms), namely no space for batching since the Kafka producer
send messages as they come in. As outlined in the related task,
this setting is very tight and not recommended in general, since
there could be pressure on the Broker side. After a chat with
Andrew Otto on the task it seems that this change shouldn't impact
MediaWiki's latency, since the Eventbus extension calls EventGate
with a deferred HTTP call.
We added new metrics in Grafana to monitor the Kafka Brokers' queues
and network threads, this change should hopefully improve those
metrics once deployed.

Bug: T338357
Change-Id: Ibf83f97e2b604833865170cf35992f52aa34c60b