thumbor: add mcrouter support
What?
Introducing mcrouter to thumbor. This patch includes just
the basic stanzas to start the migration.

Why?
Thumbor talks to memcached via nutcraker, which we have
almost retired from our infra. It is necessary we migrate
to mcrouter, which is a first class (and classy may I add)
citizen of our stack.

Nutcracker will not be removed yet, but later on after
we ensure that the migration to mcrouter is complete

Bug: T318695
Change-Id: I84622c340148b38159337463e2c64b23dc90d546