add Benthos smoke test to staging
Adding a basic "smoke test" version of the future Benthos-based edge
cache invalidator service to staging. This is in no way functional --
adding this just to help with development. This release should be
removed Soon(TM) and eventually replaced with something more complete.

Also a small change to the chart to allow for specifying the message
processing.

Bug: T324200
Change-Id: I5801438dafa576b42228b914f8b4d5b1ff2d3a5a