ml-services: add new variable in chart for s3 path
This allows the user to define the variable use_model_from_wiki.
This way we can use a model that is intended to be used in another wiki
(e.g. simplewiki and testwiki use enwiki models). This way it is excplicitly defined
that we are using another model and also we don't have redundant data in swift.
If the variable is not defined previous behavior is used.

Bug: T319170
Change-Id: Iddbf644f5bad8a0973782c853c238aff07f23bcd