knative-serving: removing default logging config
The config was never applied since the related template didn't have
the correct loop in place. Now that we have it the deployment of
the new logging config knocks down the webhook, that in turn can't
validate all yaml configs and hence the helm deployment fails.
Remove the logging config to unblock the knative's deployment.

Change-Id: I1023b1403f2cae6c9834b50410e9ace377ec5b39