mesh.certificate: Ensure the commonName is at most 64 bytes long
The FQDN of the service can easily exceed the 64 bytes limit. Switch to
using the name of the kubernetes certificate object instead (truncate
to be sure) as commonName should not be used for validation anyways.

Silent fix in mesh.certificate 1.0.0 as there are are ownly two test
users as of now.

Bug: T300033
Change-Id: I9b6739aaae96a5efcdc2a2a5b66d671ece2f1873