Use the new DataHub images built with GitLab-CI
We have now migrated the build pipeline for DataHub away from Gerrit and
Zuul to GitLab-CI, effectively completing the un-forking operation from
the datahub upstream sources.

However, the new images are hosted using a different path on our docker
registry, so we need to update this path component as well as the image
tag.

Bug: T341194
Change-Id: I6d99e43333f975a344f4e66607e80edd7154a2be