benthos: temporarily disable readiness probe
The readiness probe is failing and thus helmfile will not deploy the
chart (atomic is set => it will undo the deployment). In order to debug
this, I need the deployment to go through, so here I temporarily disable
the readiness probe.

This should be reverted (or otherwise fixed) once the problem is solved.

Bug: T324200
Change-Id: Idf8431839f6d0b44f1e7f94af389a6723e78d6b8