mediawiki: Raise syslog max_message_size
Why?
JSON messages to logstash are being truncated.
On mediawiki bare metal hosts, the rsyslog MaxMessageSize is 64k. This
setting was not ported to the mw-on-k8s rsyslog configuration.

What?
Add a configmap for a config file under /etc/rsyslog.d specifying
$MaxMessageSize.
This defaults in the charts values.yaml to 64k (as in bare-metal), but
can be overridden by the mw.logging.rsyslog_max_message_size
configuration entry in the helmfile values.yaml.

Bug: T343390
Change-Id: If656bdc1245af0e582b606475bb43799d9ad5e16