admin_ng: set target-burst-capacity to zero for knative-serving
The Activator pods can be placed between Ingress and target pods
when a surge of traffic happens, to buffer requests. Configuring
it in the right way is not easy, and it adds another layer of proxy
in the request path (that could add latency etc..).

Set target-burst-capacity to zero to limit the activator's presence
to only when pods are scaled to zero (waiting for new requests
to come before spinning up more pods).

Decrease also the number of Activator pods to two, since they will
have very little things to do after this change.

Change-Id: Ida60a6f8c34c46e5e6b3c1cad1cde02163303c2c